add: 즐겨찾기 interface 추가
FavoriteService 구현체를 직접 받는것이 아닌 interface 사용으로 변경하여 결합도를 낮춤
